requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
langchain
langchain_core
langchain-community
langchain-google-genai

langgraph
langgraph_supervisor

pydantic
typing_extensions
IPython
langsmith
dotenv
google-generativeai


fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
requests==2.31.0

